backend/requirements.txt
Django==5.2.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-dotenv==1.0.1
PyMuPDF==1.24.10
python-docx==1.1.2
groq==0.11.0
